extract logout logic
