+++ src/components/ui/user-nav.tsx
@@ -12,64 +12,16 @@ import {
   DropdownMenuShortcut,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-
-import { useUserSessionMutation } from "@/lib/api/user-sessions";
+import { useLogoutUser } from "@/lib/hooks/use-logout-user";
 import { useAuthStore } from "@/lib/providers/auth-store-provider";
-import { useCoachingRelationshipStateStore } from "@/lib/providers/coaching-relationship-state-store-provider";
-import { useCoachingSessionStateStore } from "@/lib/providers/coaching-session-state-store-provider";
-import { useOrganizationStateStore } from "@/lib/providers/organization-state-store-provider";
 import { userSessionFirstLastLettersToString } from "@/types/user-session";
-import { useRouter } from "next/navigation";
 import Link from "next/link";
 
 export function UserNav() {
-  const router = useRouter();
-  const { logout } = useAuthStore((action) => action);
   const { userSession } = useAuthStore((state) => ({
     userSession: state.userSession,
   }));
-  const { delete: deleteUserSession } = useUserSessionMutation();
-  const { resetOrganizationState } = useOrganizationStateStore(
-    (action) => action
-  );
-  const { resetCoachingRelationshipState } = useCoachingRelationshipStateStore(
-    (action) => action
-  );
-  const { resetCoachingSessionState } = useCoachingSessionStateStore(
-    (action) => action
-  );
-
-  async function logout_user() {
-    try {
-      console.trace("Resetting CoachingSessionStateStore state");
-      resetCoachingSessionState();
-
-      console.trace("Resetting CoachingRelationshipStateStore state");
-      resetCoachingRelationshipState();
-
-      console.trace("Resetting OrganizationStateStore state");
-      resetOrganizationState();
-
-      console.trace(
-        "Deleting current user session from backend: ",
-        userSession.id
-      );
-      await deleteUserSession();
-    } catch (err) {
-      console.warn("Error while logging out session: ", userSession.id, err);
-    } finally {
-      // Ensure we still log out of the frontend even if the backend request
-      // to delete the user session fails.
-      console.trace("Resetting AuthStore state");
-      logout();
-      
-      console.debug("Navigating to /");
-      await router.replace("/");
-      
-      // The backend will handle cookie deletion when the session is deleted
-      console.debug("Navigation to / completed successfully.");
-    }
-  }
+  const logoutUser = useLogoutUser();
 
   return (
     <DropdownMenu>
@@ -108,7 +60,7 @@ export function UserNav() {
           </DropdownMenuItem>
         </DropdownMenuGroup>
         <DropdownMenuSeparator />
-        <DropdownMenuItem onClick={logout_user}>
+        <DropdownMenuItem onClick={logoutUser}>
           Log out
           <DropdownMenuShortcut>⇧⌘Q</DropdownMenuShortcut>
         </DropdownMenuItem>
